Move node type out of overlay service
