+++ trin-core/src/portalnet/types/mod.rs
@@ -1,3 +1,4 @@
 pub mod content_key;
 pub mod messages;
 pub mod metric;
+pub mod node;
